Create README.md
Pay attention to the 10th line -Identity "Here we set the user group from the list of attributes"
In Select, we set the attributes that will be unloaded from AD.

All that's next are parts of the contact attributes. For example, download ecard from iphone, and you will understand what this or that attribute is called.